requirements.txt
psycopg2-binary
python-dotenv
opencv-python
numpy
torch
torchvision
ultralytics
